ordenar_y_depurar_parte_2

+++ src/main/java/ec/edu/uce/certificadorforense/infrastructure/adapters/extractors/imagen/Jfif.java
@@ -9,14 +9,11 @@ public class Jfif implements MetadataExtractor<MetadatosImagen.MetadatosImagenBu
     @Override
     public void extraer(Metadata metadata, MetadatosImagen.MetadatosImagenBuilder builder) {
         JfifDirectory dir = metadata.getFirstDirectoryOfType(JfifDirectory.class);
-
         if (dir != null) {
             Integer x = dir.getInteger(JfifDirectory.TAG_RESX);
             Integer y = dir.getInteger(JfifDirectory.TAG_RESY);
-
-            if (x != null) builder.pixelesPorUnidadX(x.longValue());
-            if (y != null) builder.pixelesPorUnidadY(y.longValue());
-
+            if (x != null) builder.pixelesPorUnidadX(x);
+            if (y != null) builder.pixelesPorUnidadY(y);
             builder.unidadFisica(dir.getDescription(JfifDirectory.TAG_UNITS));
         }
     }
